Refactor RuleSet (#477)
Refactors the RuleSet class to create AbstractRuleSet and TypedRuleSet. Normal usage is not impacted - the RuleSet class can be used as it was previously. Internally, this provides capability to create TypedRuleSets of Rules that have more properties than the ApplicationInspector rules. These properties don't affect rule logic in the processor.