WIP new Sarif Viewer to replace AI HTML format
Creating a Blazor based Sarif viewer capable of viewing Sarif reports including the AI Sarif report. The goal of this PR is to replicate the features present in the current HTML report - but in a standalone viewer written in Blazor rather than a static html report.

No ETA as of yet.